+++ ar-SA/step_10.md
@@ -1,7 +1,15 @@
 --- challenge ---
+
 ## التحدي: أنشئ موارد أخرى
+
 هل يمكنك إضافة موارد وقواعد صناعة أخرى إلى لعبتك؟
 
-
-
 --- /challenge ---
+
+***
+
+## ترجمة المجتمع المساهم 
+
+تمت ترجمة هذا المشروع بواسطة **نوره سالم الدوسري** ومراجعته بواسطة **نادية علي قاسم** 
+
+يساعدنا متطوعونا الرائعون للترجمة في منح الأطفال حول العالم فرصة تعلم البرمجية. يمكنك مساعدتنا في الوصول إلى المزيد من الأطفال عن طريق ترجمة مشاريعنا - اقرأ المزيد على [rpf.io/translate](https://rpf.io/translate).
