+++ ar-SA/step_2.md
@@ -1,28 +1,23 @@
 ## لعب اللعبة
 
-
-
-+ افتح مشروع trinket هذا: <a href="http://jumpto.cc/codecraft-go" target="_blank">jumpto.cc/codecraft-go</a>. 
++ افتح الرابط التالي للـ trinket: [trinket.io/python/025ca30b4b](http://trinket.io/python/025ca30b4b){:target="_blank"}.
 
 + استخدم مفاتيح WASD على لوحة المفاتيح لتحريك اللاعب في كل مكان في عالم اللعبة المليء بالموارد المختلفة (التراب والحشائش والماء).
-
+    
     ![screenshot](images/craft-move.png)
 
 + يمكنك الضغط على مفتاح المسافة لجمع الموارد. اجمع بعض المربعات من كل نوع من الموارد وسترى أنها أضيفت إلى المخزون.
-
+    
     ![screenshot](images/craft-pickup.png)
 
 + اضغط على مفاتيح الأرقام (1 إلى 3) لوضع مورد على الخريطة. على سبيل المثال، اضغط على 3 لتضع بعض مربعات الماء على الخريطة. ولن يمكنك القيام بذلك إلا إذا كان لديك بعض مربعات الماء في المخزون.
-
+    
     ![screenshot](images/craft-place-water.png)
 
-يمكنك صناعة عنصر بالضغط على المفتاح الذي يظهر في القائمة. والمقصود بالصناعة هو جمع العناصر الموجودة بالفعل في المخزون لإنشاء عناصر جديدة. جرِّب الضغط على المفتاح 'R' لصناعة طوبة جديدة (إذا كان لديك مربعان من التراب ومربع واحد من الماء في المخزون).
++ يمكنك صناعة عنصر بالضغط على المفتاح الذي يظهر في القائمة. والمقصود بالصناعة هو جمع العناصر الموجودة بالفعل في المخزون لإنشاء عناصر جديدة. جرِّب الضغط على المفتاح 'R' لصناعة لبنة جديدة (إذا كان لديك مربعان من التراب ومربع واحد من الماء في المخزون).
+    
+    ![screenshot](images/craft-craft-brick.png)
 
-   ![screenshot](images/craft-craft-brick.png)
-
-+ يمكنك بعد ذلك الضغط على المفتاح '4' لوضع الطوب الذي صنعتَه على الخريطة.
-
++ يمكنك بعد ذلك الضغط على المفتاح '4' لوضع اللبنة التي صنعتها على الخريطة.
+    
     ![screenshot](images/craft-place-brick.png)
-
-
-
