+++ ar-SA/step_3.md
@@ -1,10 +1,9 @@
 --- challenge ---
+
 ## التحدي: صمِّم عالمك
+
 هل يمكنك تصميم منزل بحديقة وحمام سباحة؟ ما التصميمات الأخرى التي يمكنك إنشاؤها؟
 
 ![screenshot](images/craft-build-example.png)
-
-
-
 
 --- /challenge ---
