+++ ar-SA/step_4.md
@@ -1,16 +1,15 @@
 ## تخصيص اللعبة
 
-لنعدِّل بعض المتغيرات لنغيِّر طريقة عرض اللعبة.
+لنقم بتعديل بعض المتغيرات لتغيير طريقة عرض اللعبة.
 
 + انقر فوق الملف `variables.py` لترى بعض المتغيرات التي يمكن تغييرها.
-
+    
     ![screenshot](images/craft-variables.png)
 
 + غيِّر قيمة المتغير `BACKGROUNDCOLOUR`، وانقر فوق 'Run' لترى التغيير الذي حدث في اللعبة.
-
+    
     ![screenshot](images/craft-background.png)
 
 + يمثل المتغير `MAXTILES` كمية المربعات التي يمكن أن يستوعبها المخزون من كل مورد. غيِّر هذا المتغير إذا كنت تريد تخزين أكثر (أو أقل) من 20 مربعًا من كل مورد.
-
+    
     ![screenshot](images/craft-maxtiles.png)
-
