+++ ar-SA/step_5.md
@@ -1,10 +1,9 @@
 --- challenge ---
+
 ## التحدي: غيِّر حجم عالمك
+
 هل يمكنك تغيير قيمة المتغيرين `MAPWIDTH` و`MAPHEIGHT` لتغيير حجم عالمك؟
 
-![screenshot](images/craft-mapsize.png)
-
-
-
+![لقطة شاشة](images/craft-mapsize.png)
 
 --- /challenge ---
