+++ ar-SA/step_7.md
@@ -1,14 +1,13 @@
 --- challenge ---
+
 ## التحدي: أنشئ مورد الرمل
+
 هل يمكنك إضافة مورد `SAND` إلى لعبتك؟ يمكنك استخدام الخطوات السابقة لمساعدتك.
 
 ![screenshot](images/craft-sand.png)
 
-يحتوي هذا المشروع بالفعل على صورة `sand.png`، لكن يمكنك إنشاء أو تحميل صورة من عندك إذا أردتَ.
+يحتوي هذا المشروع على صورة `sand.gif` مسبقاً، لكن يمكنك إنشاء أو تحميل صورتك الخاصة إذا أردتَ.
 
 ![screenshot](images/craft-upload.png)
-
-
-
 
 --- /challenge ---
