+++ ar-SA/step_8.md
@@ -1,42 +1,39 @@
 ## صناعة ألواح خشبية من مورد الخشب
 
-لننشئ مورد ألواح خشبية جديدًا يمكنك صناعته من مورد الخشب.
+لننشئ مورد ألواح جديد والذي من الممكن صناعته من مورد الخشب.
 
-+ أولًا، أضف متغير `PLANK` جديدًا إلى لعبتك.
-
++ أولًا، أضف متغير` PLANK` جديدًا إلى لعبتك.
+    
     ![screenshot](images/craft-plank-const.png)
 
 + أضف مورد `PLANK` جديدًا إلى لعبتك.
-
+    
     ![screenshot](images/craft-plank-resources.png)
 
 + سمِّ هذا المورد `'plank'`.
-
+    
     ![screenshot](images/craft-plank-names.png)
 
 + عيِّن صورة إلى المورد `PLANK`. يحتوي هذا المشروع بالفعل على صورة `plank.png`، لكن يمكنك إنشاء صورة من عندك إذا أردتَ.
-
+    
     ![screenshot](images/craft-plank-textures.png)
 
 + أضف عدد مربعات ألواح الخشب إلى المخزون.
-
+    
     ![screenshot](images/craft-plank-inventory.png)
 
-+ عيِّن مفتاحًا لوضع ألواح الخشب على خريطة اللعبة.
-
++ عيّن مفتاح لوضع الألواح.
+    
     ![screenshot](images/craft-plank-placekeys.png)
 
-+ حتى يمكن صناعة هذا المورد، ستحتاج إلى إنشاء قاعدة صناعة، بحيث تكون هذه القاعدة هي صناعة لوح خشبي من ثلاثة مربعات من مورد الخشب. + أضف هذه التعليمة البرمجية إلى القاموس `crafting`: 
-
++ حتى يمكن صناعة هذا المورد، ستحتاج إلى إنشاء قاعدة صناعة، بحيث تكون هذه القاعدة هي صناعة لوح خشبي من ثلاثة مربعات من مورد الخشب. أضف هذه التعليمة البرمجية إلى القاموس `crafting`.
+    
     ![screenshot](images/craft-plank-crafting.png)
 
 + وأخيرًا، ستحتاج إلى تعيين مفتاح لصناعة ألواح خشبية جديدة.
-
+    
     ![screenshot](images/craft-plank-craftkeys.png)
 
 + لتختبر مورد الألواح الخشبية الجديد، اجمع بعض مربعات الخشب ثم اصنع بعض الألواح الخشبية منها. وعندئذٍ يمكنك وضع الألواح الخشبية الجديدة في عالمك.
-
+    
     ![screenshot](images/craft-plank-test.png)
-
-
-
