+++ ar-SA/step_9.md
@@ -1,12 +1,11 @@
 --- challenge ---
+
 ## التحدي: اصنع الزجاج من الرمل
+
 هل يمكنك إنشاء مورد زجاج جديد، بحيث يمكنك صناعته من مورد الرمل؟ يمكنك اتباع الخطوات السابقة لمساعدتك.
 
 ![screenshot](images/craft-glass.png)
 
 يحتوي هذا المشروع بالفعل على صورة `glass.png` يمكنك استخدامها، أو يمكنك إنشاء صورة من عندك إذا أردتَ.
 
-
-
-
---- /challenge ---
+--- challenge ---
